requirements.txt
# Core YOLO framework and experiment utilities
# Requires Python 3.9+. To install: python3 -m venv .venv && source .venv/bin/activate && pip install -r requirements.txt
# For a reproducible install, use: pip freeze > requirements-lock.txt after installing.

# Ultralytics YOLO implementation and training/inference tooling (YOLO26-ready floor)
ultralytics>=8.4.23,<9.0.0

# PyTorch backend for model training/inference
torch>=2.2.0,<2.6.0

# TorchVision datasets, transforms, and model utilities aligned with PyTorch 2.2-2.5
torchvision>=0.17.0,<0.21.0

# Numerical operations and tensor/image array manipulation
numpy>=1.23.5,<2.0.0

# OpenCV bindings for image/video loading and preprocessing
opencv-python>=4.8.0,<5.0.0

# Tabular experiment logging and analysis
pandas>=1.5.0,<2.3.0

# Progress bars for training/evaluation loops
tqdm>=4.66.0,<5.0.0

# YAML config parsing for experiment settings
pyyaml>=6.0.0,<7.0.0
